Move resize E2E to new resize.go file
